Re-run versioneer install
