+++ .gitattributes
@@ -1,1 +1,2 @@
 pysmurf/_version.py export-subst
+python/pysmurf/_version.py export-subst
