+++ python/pysmurf/__init__.py
@@ -17,6 +17,5 @@
 # contained in the LICENSE.txt file.
 #-----------------------------------------------------------------------------
 
-from ._version import get_versions
-__version__ = get_versions()['version']
-del get_versions
+from . import _version
+__version__ = _version.get_versions()['version']
